+++ packages/plugin-store/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.0.7
 
-- [fix] get alias from userConfig object directly
+- [fix] get alias from webpack config instead of userConfig object
 
 ## 2.0.6